requirements.txt
# Core
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
httpx>=0.27.0

# Integrations
PyGithub>=2.3.0
datadog-api-client>=2.26.0
jira>=3.8.0
atlassian-python-api>=3.41.0
slack-sdk>=3.30.0
slack-bolt>=1.21.0

# Storage
# (SQLite is stdlib; upgrade to asyncpg for Postgres cloud tier)

# Dev
python-dotenv>=1.0.0
